% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_hello_world_hidden}
\alias{rcpp_hello_world_hidden}
\title{Sample Rcpp function, exported to R but hidden (not in the NAMESPACE)}
\usage{
rcpp_hello_world_hidden()
}
\value{
TRUE
}
\description{
Sample Rcpp function, exported to R but hidden (not in the NAMESPACE)
}
