% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_hello_world_internal}
\alias{rcpp_hello_world_internal}
\title{Sample Rcpp function, exported to R but internal}
\usage{
rcpp_hello_world_internal()
}
\value{
TRUE
}
\description{
Sample Rcpp function, exported to R but internal
}
\keyword{internal}
