% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_hello_world_source_call}
\alias{rcpp_hello_world_source_call}
\title{Sample Rcpp function, calling an internal function in C++}
\usage{
rcpp_hello_world_source_call()
}
\description{
Sample Rcpp function, calling an internal function in C++
}
